qsdsd

+++ client/src/app/select-animal/select-animal.component.ts
@@ -1,5 +1,6 @@
 import { Component, ElementRef, OnInit, ViewChild } from "@angular/core";
 import "bootstrap/dist/js/bootstrap.bundle";
+import { Animal } from "../../../../common/tables/Animal";
 import { CommunicationService } from "../communication.service";
 
 @Component({
@@ -16,18 +17,25 @@ export class SelectAnimalComponent implements OnInit {
   @ViewChild("button2") private button2: ElementRef;
   @ViewChild("button3") private button3: ElementRef;
   @ViewChild("button4") private button4: ElementRef;
+  @ViewChild("modification") private modification: ElementRef;
 
   public animalsId: string[];
   public cliniquesId: string[];
   public proprietairesId: string[];
+  public animalTreatments: string[];
+  public animalReceipt: string[];
   public modify: boolean;
   public treatment: boolean;
   public receipt: boolean;
+  public modifiedInput: string;
 
   public constructor(private communicationService: CommunicationService) {
     this.animalsId = [];
     this.cliniquesId = [];
     this.proprietairesId = [];
+    this.animalTreatments = [];
+    this.animalReceipt = [];
+    this.modifiedInput = "";
     this.modify = false;
     this.treatment = false;
     this.receipt = false;
@@ -39,15 +47,12 @@ export class SelectAnimalComponent implements OnInit {
 
   private getCliniquesId(): void {
     this.communicationService.getCliniquesId().subscribe((cliniquesId: string[]) => {
-      console.log(cliniquesId);
       this.cliniquesId = cliniquesId;
     });
   }
 
   public getProprietairesId(): void {
-    console.log(this.cliniqueId.nativeElement.value);
     this.communicationService.getProprietairesId(this.cliniqueId.nativeElement.value).subscribe((proprietairesId: string[]) => {
-      console.log(proprietairesId);
       this.proprietairesId = proprietairesId;
     });
   }
@@ -59,14 +64,42 @@ export class SelectAnimalComponent implements OnInit {
       console.log(this.animalsId);
     });
   }
-  
+
   public deleteAnimal(): void {
-    this.communicationService.removeAnimal(this.cliniqueId.nativeElement.value,
-                                           this.proprietaireId.nativeElement.value,
-                                           this.animalsId.nativeElement.value).subscribe((animalsId: string[]) => {
-      this.animalsId = animalsId;
-      console.log(this.animalsId);
-    });
+    const res: boolean = confirm("Etes-vous sur de vouloir supprimer cet animal?");
+    if (res) {
+      this.communicationService.removeAnimal(this.cliniqueId.nativeElement.value,
+                                             this.proprietaireId.nativeElement.value,
+                                             this.animalId.nativeElement.value).subscribe((animal: Animal[]) => {
+          console.log("Deleted!");
+        });
+    }
+  }
+
+  public getAnimalTreatments(): void {
+    this.communicationService.getAnimalTreatments(this.cliniqueId.nativeElement.value,
+                                                  this.proprietaireId.nativeElement.value,
+                                                  this.animalId.nativeElement.value).subscribe((animalTreatments: string[]) => {
+        this.animalTreatments = animalTreatments;
+      });
+  }
+
+  public getAnimalReceipt(): void {
+    this.communicationService.getAnimalReceipt(this.cliniqueId.nativeElement.value,
+                                               this.proprietaireId.nativeElement.value,
+                                               this.animalId.nativeElement.value).subscribe((animalReceipt: string[]) => {
+        this.animalReceipt = animalReceipt;
+      });
+  }
+
+  public modifyAnimalAttribute(): void {
+    this.communicationService.modifyAnimalAttribute(this.cliniqueId.nativeElement.value,
+                                                    this.proprietaireId.nativeElement.value,
+                                                    this.animalId.nativeElement.value,
+                                                    this.modification.nativeElement.value,
+                                                    this.modifiedInput).subscribe((animal: Animal[]) => {
+                                                      console.log("Modified!");
+                                                    });
   }
 
   public setButtons(): void {
@@ -89,10 +122,13 @@ export class SelectAnimalComponent implements OnInit {
         break;
       }
       case "treatment": {
+        this.getAnimalTreatments();
         this.treatment = true;
         break;
       }
       case "receipt": {
+        this.getAnimalTreatments();
+        this.getAnimalReceipt();
         this.receipt = true;
       }
     }
